services/telegram_bot/requirements.txt
python-telegram-bot
httpx
pydantic-settings
